+++ packages/box-runtime/test/fixtures/disposable-host.cjs
@@ -1,0 +1,2 @@
+"use strict";
+setInterval(() => {}, 1000);